requirements.txt
streamlit
openai>=1.6.0
python-dotenv
PyPDF2
python-docx
